backend/requirements.txt
asgiref==3.2.10
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv===0.20.0
django-filter==2.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
djoser==2.1.0
Pillow==9.2.0
flake8==4.0.1
flake8-broken-line==0.4.0
sorl-thumbnail==12.8.0
reportlab==3.5.34
requests==2.26.0
django-debug-toolbar
drf-extra-fields
django-extensions
